projetos/pubSub/requirements.txt
google-cloud-pubsub>=2.0.0
python-dotenv
